Recurse readRepo, fix race condition leading to missing commits. Closes GH-1
Previously all repository reads were kicked off in a loop, calling
readRepo with a callback which would be invoked when the last repo
was read. However this caused the commit log to be incomplete if
the last repo finished reading before all the others, clearly
incorrect.

To fix this, instead call readProject() recursively, as the readRepo
completes for each project, serializing the reads. All of the
commit log walks should now complete before the new overall git
commit log message is generated for all updated projects.